@@ -1,45 +1,50 @@
 {
-  "name": "web",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "@ironbay/dynamic": "^1.0.10",
-    "moment": "^2.24.0",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
-    "react-redux": "^6.0.1",
-    "react-router": "^5.0.0",
-    "react-router-dom": "^5.0.0",
-    "react-scripts": "2.1.8",
-    "redux": "^4.0.1",
-    "redux-thunk": "^2.3.0",
-    "uuid": "^3.3.2"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": "react-app"
-  },
-  "browserslist": [
-    ">0.2%",
-    "not dead",
-    "not ie <= 11",
-    "not op_mini all"
-  ],
-  "devDependencies": {
-    "@types/jest": "^24.0.11",
-    "@types/node": "^11.13.0",
-    "@types/react": "^16.8.13",
-    "@types/react-dom": "^16.8.3",
-    "@types/react-redux": "^7.0.6",
-    "@types/react-router": "^4.4.5",
-    "@types/react-router-dom": "^4.3.1",
-    "@types/redux": "^3.6.0",
-    "@types/uuid": "^3.4.4",
-    "typescript": "^3.4.2"
-  }
+	"name": "web",
+	"version": "0.1.0",
+	"private": true,
+	"dependencies": {
+		"@cerp/former": "^1.0.1",
+		"@cerp/syncr": "^1.1.0",
+		"@ironbay/dynamic": "^1.0.10",
+		"moment": "^2.24.0",
+		"react": "^16.11.0",
+		"react-dom": "^16.11.0",
+		"react-redux": "^6.0.1",
+		"react-router": "^5.0.0",
+		"react-router-dom": "^5.0.0",
+		"react-scripts": "2.1.8",
+		"redux": "^4.0.1",
+		"redux-thunk": "^2.3.0",
+		"uuid": "^3.3.2"
+	},
+	"scripts": {
+		"start": "react-scripts start",
+		"build": "react-scripts build",
+		"test": "react-scripts test",
+		"eject": "react-scripts eject"
+	},
+	"eslintConfig": {
+		"extends": "react-app"
+	},
+	"browserslist": [
+		">0.2%",
+		"not dead",
+		"not ie <= 11",
+		"not op_mini all"
+	],
+	"devDependencies": {
+		"@types/jest": "^24.0.11",
+		"@types/node": "^11.13.0",
+		"@types/react": "^16.9.11",
+		"@types/react-dom": "^16.9.3",
+		"@types/react-redux": "^6.0.9",
+		"@types/react-router": "^4.4.5",
+		"@types/react-router-dom": "^4.3.1",
+		"@types/redux": "^3.6.0",
+		"@types/uuid": "^3.4.4",
+		"@typescript-eslint/eslint-plugin": "^2.11.0",
+		"@typescript-eslint/parser": "^2.11.0",
+		"eslint-plugin-react": "^7.17.0",
+		"typescript": "^3.7.2"
+	}
 }
