@@ -1,5 +1,6 @@
-import Syncr from '../syncr'
-import { MergeAction, DeletesAction, QueueAction, sendServerAction, createLoginSucceed, createMerges, createDeletes } from './core'
+import Syncr from '@cerp/syncr'
+import { createLoginSucceed } from './core'
+import { connect } from 'react-redux';
 
 type Dispatch = (action: any) => any;
 type GetState = () => RootReducerState
@@ -20,22 +21,122 @@ export const createLogin = (username: string, password: string) => (dispatch: Di
 			password
 		}
 	})
-	.then((res: { token: string  }) => {
+		.then((res: { token: string }) => {
 
-		dispatch(createLoginSucceed(username, res.token, {}))
+			dispatch(createLoginSucceed(username, res.token, {}))
+		})
+}
+
+export const placeOrder = (product: Product, school: CERPSchool) => (dispatch: Dispatch, getState: GetState, syncr: Syncr) => {
+	const state = getState();
+
+	if (!state.connected) {
+		syncr.onNext('connect', () => dispatch(placeOrder(product, school)))
+		return;
+	}
+
+	syncr.send({
+		type: "PLACE_ORDER",
+		client_type: state.auth.client_type,
+		client_id: state.client_id,
+		id: state.auth.id,
+		payload: {
+			product,
+			refcode: school.refcode,
+			school_name: school.school_name,
+			school_number: school.phone_number
+		}
 	})
+		.then(res => alert("successfully placed order"))
+		.catch(err => {
+			console.error(err)
+			alert(err)
+		})
+}
+
+export const ADD_PRODUCTS = "ADD_PRODUCTS"
+export const getProducts = (filters = {}) => (dispatch: Dispatch, getState: GetState, syncr: Syncr) => {
+
+	const state = getState();
+
+	if (!state.connected) {
+		syncr.onNext('connect', () => dispatch(getProducts(filters)))
+		return;
+	}
+
+	dispatch({
+		type: "LOAD_PRODUCTS"
+	})
+
+	syncr.send({
+		type: "GET_PRODUCTS",
+		client_type: state.auth.client_type,
+		client_id: state.client_id,
+		id: state.auth.id,
+		payload: {
+			filters
+		},
+		last_sync: state.products.last_sync
+	})
+		.then((res: any) => {
+			// now dispatch an action that 'saves' these products
+
+			dispatch({
+				type: ADD_PRODUCTS,
+				products: res.products
+			})
+
+			return res
+		})
+		.catch(err => {
+			console.error(err)
+		})
 }
 
 export interface AddSchoolAction {
-	type: "ADD_SCHOOLS",
+	type: "ADD_SCHOOLS"
 	schools: { [id: string]: CERPSchool }
+}
+
+export const getSchoolProfileFromNumber = (phone_number: string) => (dispatch: Dispatch, getState: GetState, syncr: Syncr) => {
+
+	const state = getState();
+
+	if (!state.connected) {
+		syncr.onNext('connect', () => dispatch(getSchoolProfileFromNumber(phone_number)))
+		return;
+	}
+
+	syncr.send({
+		type: "GET_SCHOOL_FROM_NUMBER",
+		client_type: state.auth.client_type,
+		client_id: state.auth.id,
+		id: state.auth.id,
+		payload: {
+			phone_number
+		}
+	})
+		.then(res => {
+			console.log(res)
+			dispatch({
+				type: ADD_SCHOOLS,
+				schools: {
+					[res.id]: res.profile
+				}
+			})
+		})
+		.catch(err => {
+			console.error(err)
+			alert(err)
+		})
+
 }
 
 export const getSchoolProfiles = (school_ids: string[]) => (dispatch: Dispatch, getState: GetState, syncr: Syncr) => {
 
 	const state = getState();
 
-	if(school_ids.length > 1) {
+	if (school_ids.length > 1) {
 		alert("only do 1 at a time....")
 	}
 
@@ -48,7 +149,7 @@ export const getSchoolProfiles = (school_ids: string[]) => (dispatch: Dispatch, 
 			school_ids
 		}
 	})
-	.then(res => {
+		.then(res => {
 			console.log(res);
 			dispatch({
 				type: ADD_SCHOOLS,
@@ -66,14 +167,14 @@ export const getSchoolProfiles = (school_ids: string[]) => (dispatch: Dispatch, 
 
 export const INCOMING_PHONE_CALL = "INCOMING_PHONE_CALL";
 export interface IncomingPhoneCallAction {
-	type: "INCOMING_PHONE_CALL",
+	type: "INCOMING_PHONE_CALL"
 	number: string
 	dialed: string
 	event: string
 	unique_id: string
 }
 
-export const editSchoolProfile = (school_id: string, school: CERPSchool) => (dispatch: Dispatch, getState: GetState, syncr: Syncr)  => {
+export const editSchoolProfile = (school_id: string, school: CERPSchool) => (dispatch: Dispatch, getState: GetState, syncr: Syncr) => {
 
 	const state = getState();
 
@@ -88,12 +189,12 @@ export const editSchoolProfile = (school_id: string, school: CERPSchool) => (dis
 		id: state.auth.id,
 		last_snapshot: state.last_snapshot
 	})
-	.then(res => {
-		console.log(res)
-	})
-	.catch(err => {
-		console.error(err)
-	})
+		.then(res => {
+			console.log(res)
+		})
+		.catch(err => {
+			console.error(err)
+		})
 
 }
 
@@ -111,12 +212,12 @@ export const reserveMaskedNumber = (school_id: string) => (dispatch: Dispatch, g
 		id: state.auth.id,
 		last_snapshot: state.last_snapshot
 	})
-	.then(res => {
-		console.log(res)
-		dispatch(res)
-	})
-	.catch(err => {
-		console.error(err)
-	})
+		.then(res => {
+			console.log(res)
+			dispatch(res)
+		})
+		.catch(err => {
+			console.error(err)
+		})
 
 }
