@@ -1,6 +1,6 @@
 import { v4 } from 'uuid';
 
-export const saveDB = (db : RootReducerState) => {
+export const saveDB = (db: RootReducerState) => {
 
 	try {
 
@@ -10,7 +10,7 @@ export const saveDB = (db : RootReducerState) => {
 		saveQueue(db.queued)
 	}
 
-	catch(err) {
+	catch (err) {
 		console.error(err);
 	}
 }
@@ -24,7 +24,7 @@ export const clearDB = () => {
 
 export const loadAuth = (): RootReducerState['auth'] => {
 
-	let init_auth : RootReducerState['auth'] = {
+	const init_auth: RootReducerState['auth'] = {
 		id: undefined,
 		token: undefined,
 		client_type: "portal_call_center"
@@ -32,13 +32,13 @@ export const loadAuth = (): RootReducerState['auth'] => {
 
 	try {
 		const str = localStorage.getItem("auth")
-		if(str === null) {
+		if (str === null) {
 			return init_auth;
 		}
 
 		return JSON.parse(str);
 	}
-	catch(err) {
+	catch (err) {
 		console.error(err);
 		return init_auth;
 	}
@@ -56,11 +56,11 @@ const loadClientId = () => {
 	return client_id;
 }
 
-const loadSyncState = () : RootReducerState['sync_state'] => {
+const loadSyncState = (): RootReducerState['sync_state'] => {
 
 	const str = localStorage.getItem("sync_state");
 
-	if(str) {
+	if (str) {
 		const curr = JSON.parse(str) as RootReducerState['sync_state']
 		return curr;
 	}
@@ -69,12 +69,12 @@ const loadSyncState = () : RootReducerState['sync_state'] => {
 
 }
 
-const saveSyncState = (sync_state : RootReducerState['sync_state']) => {
+const saveSyncState = (sync_state: RootReducerState['sync_state']) => {
 
 	localStorage.setItem("sync_state", JSON.stringify(sync_state));
 }
 
-const saveQueue = (queue : RootReducerState['queued']) => {
+const saveQueue = (queue: RootReducerState['queued']) => {
 
 	localStorage.setItem("queued", JSON.stringify(queue))
 
@@ -84,7 +84,7 @@ const loadQueue = () => {
 	return JSON.parse(localStorage.getItem("queued") || "{}") as RootReducerState['queued']
 }
 
-const saveSnapshot = (last_snapshot : number) => {
+const saveSnapshot = (last_snapshot: number) => {
 
 	//@ts-ignore
 	localStorage.setItem("last_snapshot", last_snapshot);
@@ -94,7 +94,7 @@ const loadSnapshot = () => {
 	return parseInt(localStorage.getItem("last_snapshot") || "0")
 }
 
-export const loadDB = () : RootReducerState => {
+export const loadDB = (): RootReducerState => {
 
 	return {
 		client_id: loadClientId(),
@@ -105,6 +105,13 @@ export const loadDB = () : RootReducerState => {
 		connected: false,
 		sync_state: loadSyncState(),
 		active_school: undefined,
-		caller_id: undefined
+		caller_id: undefined,
+		products: {
+			last_sync: 0,
+			loading: true,
+			db: {
+
+			}
+		}
 	}
 }
