@@ -1,4 +1,4 @@
-export default function(ms : number) {
+export default (ms: number) => {
 	return new Promise((resolve) => {
 		setTimeout(() => resolve(ms), ms)
 	})
