@@ -29,11 +29,11 @@ defmodule EdMarkaz.ActionHandler.CallCenter do
 	end
 
 	def handle_action(%{"type" => "GET_SCHOOL_PROFILES", "payload" => payload}, %{id: id, client_id: client_id} = state) do
-		
+
 		ids = Map.get(payload, "school_ids", [])
 
 		or_str = Stream.with_index(ids, 1)
-			|> Enum.map(fn {_, i} -> 
+			|> Enum.map(fn {_, i} ->
 				"id=$#{i}"
 			end)
 			|> Enum.join(" OR ")
@@ -53,10 +53,10 @@ defmodule EdMarkaz.ActionHandler.CallCenter do
 
 	def handle_action(%{"type" => "SAVE_SCHOOL", "payload" => %{"school_id" => school_id, "school" => school}, "last_snapshot" => last_sync_date}, %{id: id, client_id: client_id} = state) do
 
-		case Postgrex.query(EdMarkaz.DB, "UPDATE platform_schools set db = $1 where id=$2", [school, school_id]) do 
-			{:ok, res} -> 
+		case Postgrex.query(EdMarkaz.DB, "UPDATE platform_schools set db = $1 where id=$2", [school, school_id]) do
+			{:ok, res} ->
 				{:reply, succeed(), %{id: id, client_id: client_id}}
-			{:error, msg} -> 
+			{:error, msg} ->
 				IO.inspect msg
 				{:reply, fail(), state}
 		end
@@ -70,9 +70,66 @@ defmodule EdMarkaz.ActionHandler.CallCenter do
 			{:ok, res} ->
 				[[ db ]] = res.rows
 				{:reply, succeed(db), state}
-			{:error, err} -> 
+			{:error, err} ->
 				IO.inspect err
 				{:reply, fail("db error"), state}
+		end
+	end
+
+	def handle_action(%{"type" => "GET_SCHOOL_FROM_NUMBER", "payload" => %{"phone_number" => phone_number }}, %{id: id, client_id: client_id} = state) do
+
+		case EdMarkaz.School.get_profile(phone_number) do
+			{:ok, school_id, profile} ->
+				{:reply, succeed(%{ "profile" => profile, "id" => school_id }), state}
+			{:error, message} ->
+				{:reply, fail(message), state}
+		end
+
+	end
+
+	def handle_action(%{"type" => "PLACE_ORDER", "payload" => %{"product" => product, "refcode" => refcode, "school_name" => school_name, "school_number" => school_number }}, %{client_id: client_id, id: id} = state) do
+
+		IO.puts "handling order"
+		product_name = Map.get(product, "title")
+		supplier_id = Map.get(product, "supplier_id")
+
+		start_supplier(supplier_id)
+		EdMarkaz.Supplier.place_order(supplier_id, product, refcode, client_id)
+		spawn fn ->
+			EdMarkaz.Slack.send_alert("#{school_name} placed order for #{product_name} by #{supplier_id}. Their number is #{school_number}", "#platform-orders")
+		end
+
+		spawn fn ->
+			EdMarkaz.Contegris.send_sms(id, "You have requested information for #{product_name} and will be contacted soon with more information.")
+		end
+
+		{:reply, succeed(), state}
+
+	end
+
+	def handle_action(%{"type" => "GET_PRODUCTS", "last_sync" => last_sync}, state) do
+
+		dt = DateTime.from_unix!(last_sync, :millisecond)
+
+		case Postgrex.query(EdMarkaz.DB, "
+			SELECT
+				p.id,
+				p.supplier_id,
+				p.product,
+				p.sync_time,
+				s.sync_state->'profile'
+			FROM products p LEFT JOIN suppliers s ON p.supplier_id = s.id
+			WHERE extract(epoch from sync_time) > $1 ", [0]) do
+			{:ok, resp} ->
+				mapped = resp.rows
+					|> Enum.map(fn [id, supplier_id, product, sync_time, supplier_profile] -> {id, Map.put(product, "supplier_profile", supplier_profile)} end)
+					|> Enum.into(%{})
+
+				{:reply, succeed(%{products: mapped}), state}
+			{:error, err} ->
+				IO.puts "error getting product"
+				IO.inspect err
+				{:reply, fail("error getting products"), state}
 		end
 	end
 
